integrate food search API in food_search_page.dart

+++ lib/app/models/branded_food_model/branded_food_model.dart
@@ -12,7 +12,7 @@ class BrandedFoodModel with _$BrandedFoodModel {
     @JsonKey(name: 'brand_name') String? brandName,
     @JsonKey(name: 'food_name') String? foodName,
     @JsonKey(name: 'nix_item_id') String? nixItemId,
-    @JsonKey(name: 'nf_calories') int? calorieKcal,
+    @JsonKey(name: 'nf_calories') double? calorieKcal,
     @JsonKey(name: 'serving_qty') int? servingQty,
     @JsonKey(name: 'serving_unit') String? servingUnit,
   }) = _BrandedFoodModel;
